requirements.txt
PyYAML==6.0
Requests==2.30.0
